* XACT3 and XAudio2 APIs are canceled.

+++ trunk/Libraries/Xtro.MDX/CallbacksXACT3.h
@@ -5,7 +5,7 @@ private:
 	{
 	}
 public:
-	delegate int GetOverlappedResult(IntPtr, /*Overlapped%,*/ unsigned long%, bool);
+	delegate int GetOverlappedResult(IntPtr, /*Overlapped%,*/ unsigned int%, bool);
 	// http://msdn.microsoft.com/en-us/library/367eeye0.aspx
 	// struct OVERLAPPED
 };
